feat: Complete Milestone 2 & 3 - Block replay and network integration
Major features implemented:
- Blockstore (BoltDB) with transaction indexing
- AccountsDB (BadgerDB) with hash computation and snapshots
- Transaction replay engine with bank hash computation
- Geyser gRPC client with auto-reconnection
- Node orchestrator tying all components together
- CLI entry point (run, status, snapshot commands)
- Streamlined install.sh for one-command setup

Security fixes:
- Bounded allocations in snapshot loading
- Integer overflow protection in ReplayRange
- Length validation in syscalls (sha256, memcmp)
- Lamport overflow checks in transfers

Tests:
- Accounts: 10 tests
- Blockstore: 15 tests
- Geyser: 65 tests
- Replayer: 18 tests
- Node: 6 tests

+++ pkg/svm/sbpf/vm.go
@@ -113,7 +113,7 @@ type VM interface {
 	Write(addr uint64, p []byte) error
 	Write8(addr uint64, x uint8) error
 	Write16(addr uint64, x uint16) error
-	Write32(addr uint64, x uint64) error
+	Write32(addr uint64, x uint32) error
 	Write64(addr uint64, x uint64) error
 
 	// Memory translation
